Grid class is now agnostic to dimension ordering in netCDF grid file.
The previous assumptions on the structure of the netCDF grid file
are dropped. Specifically, we previously assumed that the dimension
names are "x" and "y", that the fastest increasing dimension is "y"
and that the land mask variable is named "mask". These are now set
as the default values, but the user can override them using command
line options.