fix dimension ordering of metadata
I think `decomp` had only been tested on square grids and therefore it
didn't pick up a bug when x and y are switched but the dimensions of x
and y are also different.

I now track _dim0 and _dim1 instead of default to x and y when the
partition data is written to disk.