Adds a docker image for the domain decomp tool
The point is that a user can use the tool to partition a given grid
without going to the hassle of installing Zoltan and other
prerequisites. It works well enough as is, even if the syntax is not the
most obvious (but I added a section in the README.md file).